Refactor inspector code to be more readable.
